fix(build-audit): refactor schema locations
