+++ libs/feature/audit-builder/src/lib/audit-builder/audit-globals.component.ts
@@ -7,7 +7,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { RxIf } from '@rx-angular/template/if';
 import { RxFor } from '@rx-angular/template/for';
-import { DEVICE_TYPE } from '../schema/audit-builder.constants';
+import { DEVICE_TYPE } from '../schema/audit.constants';
 import { ToTitleCasePipe } from '../utils/toTitleCase.pipe';
 import { AuditBuilderService } from './audit-builder.service';
